# backend/api/swagger/swagger.yaml
swagger: "2.0"
info: 
  version: "1.0.0"
  title: Toxic Raccoon API
host: localhost:4000
basePath: /api
schemes:
  - "https"
  - "http"
paths:
...
  /SENDMAIL:
    post:
      tags:
      - "Email send"
      summary: "Send mail with the ordering infos"
      description: ""
      operationId: "SendMail"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Order detailse, shipping & billing info"
        required: true
        schema:
          $ref: "#/definitions/Email"
      responses:
        "405":
          description: "Invalid input"
definitions:
...
  Email:
    type: "object"
    properties:
      bagTotalPrice: 
        type: "string"
        example: "800000"
      currentBag:
        type: "array"
        items:
          type: "object"
          example: 
            COMPONENT_LIST: []
            BUILD_NAME: "The Coon."
            USER_NAME: "Toxic Raccoon"
            BUILD_PRICE: 556848
            BUILD_ID: "f5ef6389-8a7e-476b-bf2b-c71a70ae0773"
      shippingDetailes:
        type: "object"
        example: "address"
      billingDetailes:
        type: "object"
        example: "address"
      takeOverType:
        type: "string"
        example: "shipping"
      paymentOption:
        type: "string"
        example: "in cash"
      additionalInfo:
        type: "string"
        example: "house owner name"